projeto-agno/pyproject.toml
[project]
name = "projeto-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.4.8",
    "google-genai>=1.61.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
    "yfinance>=1.1.0",
]
